requirements.txt
PyMySQL
httpx[http2]
pyquery
python-dotenv
sqlalchemy
zstandard
